feat(test): better error printing for invalid tree transforms
It's important to know when your transformations are wrong and break the equality of the before and after trees. This can be hard to spot in the output of thousands of problems, so we assert that the evaluated value of the expression doesn't change between rule applications. When the value does change, we now output the entire problem history so you can scrutinize and determine which operation broke the tree.