refactor(model): remove episode long RNN state tracking
R2D2 and friends use this to great effect, but either I don't train long enough or I messed up the implementation. Remove it for now to simplify model usage, and let LSTM only operate on the current batch of observations (which are a rolling window of timesteps)

BREAKING CHANGE: this removes long-term RNN state tracking across episodes. Tracking the state was a significant amount of code and it wasn't clear that it made the model substantially better at any given task.

The overhead associated with keeping lots of hidden states in memory and calculating state histories was not insignificant on CPU training setups as well.