refactor(worker): support retraining at end of episode
For the A3C agent we train on bootstrapped rewards throughout the episode, but at the end we get a large reward signal that ripples back through the earlier rewards.

After the episode is done, we can use the known terminal values to retrain on window_size batches of the episode.

I'm not sure if this will be terrible or not, but I think of it like training once on the estimates, and then studying the final result to fix any mistakes. TODO: :shrug: